% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3_bar.R
\name{d3_bar}
\alias{d3_bar}
\title{Bar plot D3}
\usage{
d3_bar(data, x_var, y_var = NULL, title = NULL, x_lab = NULL,
  y_lab = NULL, title_size = 18, color = "black", border = NULL,
  opacity = 1)
}
\arguments{
\item{data}{A Dataset to use for plotting. Should be a dataframe}

\item{x_var}{Attribute to place on the x axis}

\item{y_var}{Attribute for the y axis. If *y_var* is not specified or NULL, the *x_var* will be counted}

\item{title}{Title for the plot, placed above plot}

\item{x_lab}{X axis label. Default is *str_to_title(x_var)*}

\item{y_lab}{Y axis label. Default is *str_to_title(y_var)*}

\item{title_size}{Text size of the title}

\item{color}{Color of the bars. This value can be a hex color, color name, or the column name from the dataframe}

\item{border}{Color of the border of the bars}

\item{opacity}{transparency level}
}
\value{
An HTML widget that displays a bar plot made with D3.js
}
\description{
Bar plot D3
}
\examples{
# not run:
# d3_bar(mtcars, am)
}
