% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3_gadget.R
\name{d3_gadget}
\alias{d3_gadget}
\title{Interactive data exploration}
\usage{
d3_gadget()
}
\value{
Shiny Gadget that uses the D3 functions
}
\description{
Interactive data exploration
}
\examples{
# not run:
# d3_gadget()
}
