% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3_histogram.R
\name{d3_histogram}
\alias{d3_histogram}
\title{Bar plot made with D3}
\usage{
d3_histogram(data, x_var, title = NULL, x_lab = NULL, y_lab = NULL,
  color = "black", border = NULL, opacity = 1,
  transition_duration = 0)
}
\arguments{
\item{data}{A Dataset to use for plotting. Should be a dataframe}

\item{x_var}{Attribute to place on the x axis}

\item{title}{Title for the plot, placed above plot}

\item{x_lab}{X axis label. Default is *x_var's* object name}

\item{y_lab}{Y axis label. Default is "Frequency"}

\item{color}{Color of the bars. This value can be a hex color, color name, or the column name from the dataframe}

\item{border}{Color of the border of the bars}

\item{opacity}{transparency level}

\item{transition_duration}{Duration for the bars to transtion}
}
\value{
An HTML widget that displays a bar plot made with D3.js
}
\description{
Bar plot made with D3
}
\examples{
# not run:
# d3_histogram(mtcars, mpg)
}
